requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
appdirs==1.4.4
arabert==1.0.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asteroid-filterbanks==0.4.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.5
asyncio==3.4.3
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
black==21.4b2
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
camel_tools==1.5.5
catalogue==2.0.10
catboost==1.2.7
certifi==2025.8.3
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
chromadb==1.0.15
cleo==2.1.0
click==8.1.7
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
crashtest==0.4.1
cryptography==45.0.5
cycler==0.12.1
cymem==2.0.11
datasets==3.5.0
debugpy==1.8.1
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@82a57ce0b70057685962b352535147d9a8118578
dill==0.3.8
distlib==0.3.9
distro==1.9.0
docopt==0.6.2
docstring_parser==0.16
docx==0.2.4
docx2txt==0.9
dulwich==0.22.8
durationpy==0.10
editdistance==0.8.1
einops==0.8.1
emoji==1.4.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.0.1
faiss-cpu==1.11.0.post1
farasapy==0.0.14
fastjsonschema==2.21.1
filelock==3.16.1
filetype==1.2.0
findpython==0.6.3
Flask==3.0.3
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.6.0
gdown==5.2.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.170.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==3.2.0
h5py==3.12.1
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.29.3
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==5.2.0
HyperPyYAML==1.2.2
idna==3.10
imageio==2.36.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
installer==0.7.0
iopath==0.1.9
ipykernel==6.29.4
ipython==8.25.0
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jedi==0.19.1
Jinja2==3.1.4
jiter==0.10.0
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.9.2
keyring==25.6.0
kiwisolver==1.4.7
kubernetes==33.1.0
langchain==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.9
langcodes==3.5.0
langsmith==0.4.13
language_data==1.3.0
layoutparser==0.3.4
libclang==18.1.1
lightgbm==4.6.0
lightning==2.5.1
lightning-utilities==0.14.3
litellm==1.74.4
llvmlite==0.44.0
loguru==0.7.3
lxml==5.3.1
Mako==1.3.9
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mcp==1.11.0
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
mmh3==5.1.0
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.1
muddler==0.1.3
multidict==6.2.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.32.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
oauthlib==3.3.1
omegaconf==2.3.0
omniparse==0.0.3
onnxruntime==1.22.1
openai==1.97.0
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opt_einsum==3.4.0
optree==0.13.1
optuna==4.2.1
orjson==3.11.1
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pbs-installer==2025.7.12
pdf2image==1.17.0
pdf2img==0.1.2
pdfminer.six==20250506
pdfplumber==0.11.7
Pillow==9.5.0
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
pkginfo==1.12.1.2
platformdirs==4.2.2
plotly==6.0.1
poetry==2.1.3
poetry-core==2.1.3
portalocker==2.7.0
posthog==5.4.0
preshed==3.0.9
primePy==1.3
prometheus_client==0.21.1
prompt_toolkit==3.0.46
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
pure-eval==0.2.2
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
PyArabic==0.6.15
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycocotools==2.0.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.1
Pygments==2.18.0
pyjokes==0.8.3
PyJWT==2.10.1
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyresparser==1.0.6
pyrsistent==0.20.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-lightning==2.5.1
pytorch-metric-learning==2.8.1
pytz==2024.2
pywin32==311
pywin32-ctypes==0.2.3
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.0.3
RapidFuzz==3.13.0
rdkit==2025.3.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
semver==3.0.4
Send2Trash==1.8.3
sentence-transformers==5.0.0
sentencepiece==0.2.0
setuptools==75.4.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.3
SQLAlchemy==2.0.40
srsly==2.5.1
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.47.1
streamlit==1.45.0
swarms==7.9.9
swarms-memory==0.1.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.19.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
tf_keras==2.19.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.13.3
torch==2.6.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.6.0
torchmetrics==1.7.1
torchvision==0.21.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.43.4
trove-classifiers==2025.5.9.12
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.31.2
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.0
whisper==1.1.10
win32_setctime==1.2.0
wrapt==1.16.0
xgboost==3.0.0
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.3
zipp==3.23.0
zstandard==0.23.0
